LSF submission with same interface as SSH and local submission

See merge request one-touch-pipeline/weskit/api!83

+++ weskit/classes/executor/LocalExecutor.py
@@ -5,8 +5,6 @@
 #      https://gitlab.com/one-touch-pipeline/weskit/api/-/blob/master/LICENSE
 #
 #  Authors: The WESkit Team
-from __future__ import annotations
-
 import logging
 import subprocess  # nosec B603
 from datetime import datetime
@@ -16,7 +14,6 @@ from typing import Optional
 from builtins import int, super, open, property
 
 import weskit.classes.executor.Executor as base
-from weskit.classes.executor.ExecutorException import ExecutorException
 from weskit.classes.ShellCommand import ShellCommand
 
 logger = logging.getLogger(__name__)
@@ -83,26 +80,29 @@ class LocalExecutor(base.Executor):
 
     def execute(self,
                 command: ShellCommand,
-                stdout_file: Optional[PathLike] = None,
-                stderr_file: Optional[PathLike] = None,
-                stdin_file: Optional[PathLike] = None,
-                settings: Optional[base.ExecutionSettings] = None) \
+                stdout_file: Optional[base.FileRepr] = None,
+                stderr_file: Optional[base.FileRepr] = None,
+                stdin_file: Optional[base.FileRepr] = None,
+                settings: Optional[base.ExecutionSettings] = None,
+                **kwargs) \
             -> base.ExecutedProcess:
         """
         Execute the process in the background and return the process ID.
+
+        Note, the LocalExecutor supports processing of IOBase objects as stderr, stdout, stdin.
         """
-        if stdin_file is not None:
+        if isinstance(stdin_file, PathLike):
             stdin = open(stdin_file, "r")
         else:
-            stdin = None
-        if stderr_file is not None:
+            stdin = stdin_file
+        if isinstance(stderr_file, PathLike):
             stderr = open(stderr_file, "a")
         else:
-            stderr = None
-        if stdout_file is not None:
+            stderr = stdin_file
+        if isinstance(stdout_file, PathLike):
             stdout = open(stdout_file, "a")
         else:
-            stdout = None
+            stdout = stdout_file
 
         start_time = datetime.now()
 
@@ -118,39 +118,68 @@ class LocalExecutor(base.Executor):
                                    stderr=stderr,
                                    stdin=stdin,
                                    env=command.environment,
-                                   shell=False)
+                                   shell=False,
+                                   **kwargs)
             logger.debug(f"Started process {process.pid}: {command.command}")
 
-            return base.ExecutedProcess(command=command,
-                                        executor=self,
+            return base.ExecutedProcess(executor=self,
                                         process_handle=process,
                                         pre_result=base.
-                                        CommandResult(id=base.ProcessId(process.pid),
+                                        CommandResult(command=command,
+                                                      id=base.ProcessId(process.pid),
                                                       stdout_file=stdout_file,
                                                       stderr_file=stderr_file,
                                                       stdin_file=stdin_file,
                                                       run_status=base.RunStatus(),
                                                       start_time=start_time))
         except FileNotFoundError as e:
-            raise ExecutorException(f"Invalid working directory '{command.workdir}", e)
+            # The other executors recognize inaccessible working directories or missing commands
+            # only after the wait_for(). We emulate this behaviour here such that all executors
+            # behave identically with respect to these problems.
+            if e.strerror == f"No such file or directory: {command.workdir.__repr__()}":
+                # cd /dir/does/not/exist: exit code 1
+                exit_code = 1
+            else:
+                # Command not executable: exit code 127
+                exit_code = 127
+
+            return base.ExecutedProcess(executor=self,
+                                        process_handle=None,
+                                        pre_result=base.
+                                        CommandResult(command=command,
+                                                      id=base.ProcessId(None),
+                                                      stdout_file=stdout_file,
+                                                      stderr_file=stderr_file,
+                                                      stdin_file=stdin_file,
+                                                      run_status=base.RunStatus(
+                                                          exit_code, message=e.strerror),
+                                                      start_time=start_time))
 
     def get_status(self, process: base.ExecutedProcess) -> base.RunStatus:
-        return base.RunStatus(process.handle.poll())
+        if process.handle is None:
+            return process.result.status
+        else:
+            return base.RunStatus(process.handle.poll())
 
     def update_process(self, process: base.ExecutedProcess) -> base.ExecutedProcess:
         """
         Update the the executed process, if possible.
         """
-        result = process.result
-        return_code = process.handle.poll()
-        if return_code is not None:
-            result.status = base.RunStatus(return_code)
-            result.end_time = datetime.now()
-            process.result = result
-        return process
+        if process.handle is None:
+            # There is no process handle. We cannot update the result.
+            return process
+        else:
+            result = process.result
+            return_code = process.handle.poll()
+            if return_code is not None:
+                result.status = base.RunStatus(return_code)
+                result.end_time = datetime.now()
+                process.result = result
+            return process
 
     def kill(self, process: base.ExecutedProcess):
-        process.handle.terminate()
+        if process.handle is not None:
+            process.handle.terminate()
         # TODO The API says little. The code says it uses os.kill for Unix. More recherche needed.
 
     def wait_for(self, process: base.ExecutedProcess, timeout: Optional[float] = None)\
